[ZIL Kinds] move lwb-specific types, constants, and zilog_t into zil_lwb.h
