+++ include/sys/zil.h
@@ -68,33 +68,8 @@ typedef struct zil_header {
 	uint64_t zh_pad[3];
 } zil_header_t;
 
-/*
- * zh_flags bit settings
- */
-#define	ZIL_REPLAY_NEEDED	0x1	/* replay needed - internal only */
-#define	ZIL_CLAIM_LR_SEQ_VALID	0x2	/* zh_claim_lr_seq field is valid */
 
-/*
- * Log block chaining.
- *
- * Log blocks are chained together. Originally they were chained at the
- * end of the block. For performance reasons the chain was moved to the
- * beginning of the block which allows writes for only the data being used.
- * The older position is supported for backwards compatibility.
- *
- * The zio_eck_t contains a zec_cksum which for the intent log is
- * the sequence number of this log block. A seq of 0 is invalid.
- * The zec_cksum is checked by the SPA against the sequence
- * number passed in the blk_cksum field of the blkptr_t
- */
-typedef struct zil_chain {
-	uint64_t zc_pad;
-	blkptr_t zc_next_blk;	/* next block in chain */
-	uint64_t zc_nused;	/* bytes in log block used */
-	zio_eck_t zc_eck;	/* block trailer */
-} zil_chain_t;
 
-#define	ZIL_MIN_BLKSZ	4096ULL
 
 /*
  * ziltest is by and large an ugly hack, but very useful in
@@ -104,14 +79,6 @@ typedef struct zil_chain {
  * We subtract TXG_CONCURRENT_STATES to allow for common code.
  */
 #define	ZILTEST_TXG (UINT64_MAX - TXG_CONCURRENT_STATES)
-
-/*
- * The words of a log block checksum.
- */
-#define	ZIL_ZC_GUID_0	0
-#define	ZIL_ZC_GUID_1	1
-#define	ZIL_ZC_OBJSET	2
-#define	ZIL_ZC_SEQ	3
 
 typedef enum zil_create {
 	Z_FILE,
@@ -404,64 +371,6 @@ typedef struct itx {
 	/* followed by type-specific part of lr_xx_t and its immediate data */
 } itx_t;
 
-/*
- * Used for zil kstat.
- */
-typedef struct zil_stats {
-	/*
-	 * Number of times a ZIL commit (e.g. fsync) has been requested.
-	 */
-	kstat_named_t zil_commit_count;
-
-	/*
-	 * Number of times the ZIL has been flushed to stable storage.
-	 * This is less than zil_commit_count when commits are "merged"
-	 * (see the documentation above zil_commit()).
-	 */
-	kstat_named_t zil_commit_writer_count;
-
-	/*
-	 * Number of transactions (reads, writes, renames, etc.)
-	 * that have been committed.
-	 */
-	kstat_named_t zil_itx_count;
-
-	/*
-	 * See the documentation for itx_wr_state_t above.
-	 * Note that "bytes" accumulates the length of the transactions
-	 * (i.e. data), not the actual log record sizes.
-	 */
-	kstat_named_t zil_itx_indirect_count;
-	kstat_named_t zil_itx_indirect_bytes;
-	kstat_named_t zil_itx_copied_count;
-	kstat_named_t zil_itx_copied_bytes;
-	kstat_named_t zil_itx_needcopy_count;
-	kstat_named_t zil_itx_needcopy_bytes;
-
-	/*
-	 * Transactions which have been allocated to the "normal"
-	 * (i.e. not slog) storage pool. Note that "bytes" accumulate
-	 * the actual log record sizes - which do not include the actual
-	 * data in case of indirect writes.
-	 */
-	kstat_named_t zil_itx_metaslab_normal_count;
-	kstat_named_t zil_itx_metaslab_normal_bytes;
-
-	/*
-	 * Transactions which have been allocated to the "slog" storage pool.
-	 * If there are no separate log devices, this is the same as the
-	 * "normal" pool.
-	 */
-	kstat_named_t zil_itx_metaslab_slog_count;
-	kstat_named_t zil_itx_metaslab_slog_bytes;
-} zil_stats_t;
-
-extern zil_stats_t zil_stats;
-
-#define	ZIL_STAT_INCR(stat, val) \
-    atomic_add_64(&zil_stats.stat.value.ui64, (val));
-#define	ZIL_STAT_BUMP(stat) \
-    ZIL_STAT_INCR(stat, 1);
 
 typedef int zil_replay_func_t(void *arg1, void *arg2, boolean_t byteswap);
 typedef int zil_get_data_t(void *arg, uint64_t arg2, lr_write_t *lr, char *dbuf,
@@ -516,8 +425,6 @@ extern void	zil_clean(zilog_t *zilog, uint64_t synced_txg);
 extern int	zil_suspend(const char *osname, void **cookiep);
 extern void	zil_resume(void *cookie);
 
-extern void	zil_lwb_add_block(struct lwb *lwb, const blkptr_t *bp);
-extern void	zil_lwb_add_txg(struct lwb *lwb, uint64_t txg);
 extern int	zil_bp_tree_add(zilog_t *zilog, const blkptr_t *bp);
 
 extern void	zil_set_sync(zilog_t *zilog, uint64_t syncval);
