+++ module/os/linux/zfs/zvol_os.c
@@ -38,6 +38,7 @@
 #include <sys/zvol.h>
 #include <sys/zvol_impl.h>
 #include <sys/zil_impl.h>
+#include <sys/zil_lwb.h>
 
 #include <linux/blkdev_compat.h>
 #include <linux/task_io_accounting_ops.h>
