+++ module/zfs/dsl_scan.c
@@ -45,6 +45,7 @@
 #include <sys/spa_impl.h>
 #include <sys/vdev_impl.h>
 #include <sys/zil.h>
+#include <sys/zil_lwb.h>
 #include <sys/zio_checksum.h>
 #include <sys/ddt.h>
 #include <sys/sa.h>
